move meta-header of the view sub-module to its folder and update the range meta-header
